backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Configuration Management
pydantic<2.0.0
python-dotenv==1.0.0

# Database
pymongo==4.5.0
motor==3.3.1  # Async MongoDB driver

# Vector Database
pymilvus==2.3.1

# Search Engine
elasticsearch==8.9.0

# Cache
redis==5.0.0

# Machine Learning & NLP
torch==2.0.1
transformers==4.33.2
sentence-transformers==2.2.2
langchain==0.0.292
langchain-community==0.0.1

# Medical NLP Models
# These will be downloaded automatically by transformers
# dmis-lab/biobert-base-cased-v1.1
# microsoft/BiomedNLP-PubMedBERT-base-uncased-abstract
# emilyalsentzer/Bio_ClinicalBERT

# Text Processing
nltk==3.8.1
spacy==3.6.1

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Logging & Monitoring
structlog==23.1.0

# Development & Testing
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0

# Production
gunicorn==21.2.0
uvicorn==0.23.2